Dependency updates and navigation safe args moved to precompiled script
Cleaned up root build.gradle.kts.
