+++ buildSrc/build.gradle.kts
@@ -15,5 +15,7 @@ dependencies {
     implementation(libs.gradle.api)
     implementation(libs.kotlin.gradle.plugin)
 
+    implementation(libs.navigation.safe.args.plugin)
+    implementation(libs.gradle.versions.plugin)
     implementation(libs.detekt)
 }
